Move worker outside klee_web. Export pythonpath to make absolute imports from src possible.
