+++ .gitignore
@@ -18,3 +18,4 @@ src/klee_web/geoip/
 src/geoip/
 ubuntu-xenial*.log
 .vscode/
+note*.txt
