+++ provisioning/roles/common/templates/environment_dev.j2
@@ -10,3 +10,4 @@ export DB_HOST={{ db_host }}
 export DB_PORT={{ db_port }}
 export DB_NAME={{ db_name }}
 export DEVELOPMENT=1
+export PYTHONPATH="/titb/"
