+++ provisioning/roles/common/templates/environment_prod.j2
@@ -23,3 +23,4 @@ export DB_PASSWORD={{ db_password }}
 export DB_HOST={{ db_host }}
 export DB_PORT={{ db_port }}
 export DB_NAME={{ db_name }}
+export PYTHONPATH="/titb/"
