+++ src/klee_web/api/views.py
@@ -1,4 +1,5 @@
 import datetime
+
 from django.urls import reverse
 
 from rest_framework import viewsets, status
@@ -14,8 +15,8 @@ from api.permissions import IsOwnerOrReadOnly
 from api.serializers import ProjectSerializer, FileSerializer
 from frontend.models import Project, File, Task
 
-from worker.worker import submit_code
-from worker.worker_config import WorkerConfig
+from src.worker.worker import submit_code
+from src.worker.worker_config import WorkerConfig
 
 
 class ProjectViewSet(viewsets.ModelViewSet):
