+++ src/klee_web/tests/python_tests/testcase_generate.py
@@ -1,6 +1,6 @@
 import os
 
-from worker.runner import WorkerRunner
+from src.worker.runner import WorkerRunner
 
 BASE_DIR = os.path.dirname(os.path.realpath(__file__))
