+++ src/worker/processor/coverage.py
@@ -2,7 +2,7 @@ import os
 import glob
 
 from gcovparse import gcovparse
-from worker.processor.base import BaseProcessor
+from src.worker.processor.base import BaseProcessor
 
 
 class CoverageProcessor(BaseProcessor):
