+++ src/worker/processor/klee_run.py
@@ -1,4 +1,4 @@
-from worker.processor.base import BaseProcessor
+from src.worker.processor.base import BaseProcessor
 
 
 class KleeRunProcessor(BaseProcessor):
