+++ src/worker/tests/fixtures/invalid_syntax/expected.stdout
