+++ src/worker/tests/fixtures/invalid_syntax/input.c
