+++ src/worker/tests/fixtures/simple/expected.stdout
