+++ src/worker/tests/fixtures/simple/input.c
