+++ src/worker/tests/fixtures/simple_unicode/expected.stdout
