+++ src/worker/tests/fixtures/simple_unicode/input.c
