+++ src/worker/tests/fixtures/symargs/expected.stdout
