+++ src/worker/tests/fixtures/symargs/input.c
