+++ src/worker/tests/fixtures/symin/expected.stdout
