+++ src/worker/tests/fixtures/symin/input.c
